Add files via upload
This system has capable of diagnosing respiratory illnesses and decide if someone is having a cold or flu. The system requests the user to enter the symptoms as input and the system diagnosis accurately even if the user entered any synonyms of the symptoms. Then it checks if the symptom is flu symptom or cold symptom to determine which counter that need to increase flu or cold counter, it does the same step with all symptoms. Also, if the user enters less than 3 symptoms, it asked to enter more symptoms to diagnose their state. Finally, it compares between flu and cold counter to print the diagnosis result.